Gets IProblems from parser, fixed issue with comments
Previously calls to getProblems() didn't return any problems, despite SO examples that said that code SHOULD do so. The current code does somehow. I also resolved the issues with comments - we have to pass the source string to copy, as setting the IDocument to the CompilationUnit.toString is where we lost comments. Commented out imports are removed with their following node, but this expected behaviour as parsers associate comments with nodes like this.